Issue #30 - Factor out CSS for Classic skin toolbar button into separate file
For the Classic skin, the CSS for the calendar toolbar button needs to be
included everywhere, whereas currently, caldavzap.css only gets included when
actually loading the calendar view. This means that at the moment, most of the
time that button is displaying the wrong icon.

To be on the safe side, we factor out the CSS code for the button icon into a
separate file that gets included permanently, and continue loading the rest of
the CSS styles only in the calendar view itself.